refactor teams
